re-design project-structure
