--- App/Screens/SettingsScreen.js
@@ -1,12 +1,0 @@
-import React, { Component } from 'react';
-import {  View, Text, } from 'react-native';
-
-export default class componentName extends Component {
-  render() {
-    return (
-      <View>
-        
-      </View>
-    );
-  }
-}
